Util, gPESS: Use generic cast to full matrix in qr_sites
To get the full q matrix from the qr_decomposition we have to
cast q to a matrix. This commit strips the dimension of the
tensor type and uses the resulting type to construct a Matrix type
which is used to cast q into a full matrix.

The interface to get the basetype `T.name.wrapper` is a Julia
internal, but it is supposed to get a proper interface by the name
`basetype`. See: https://github.com/JuliaLang/julia/issues/35543